refactor(core): split VAD into config, silero, regions, segmenter
Move SileroVad, region math, the file segmenter, and the streaming
endpointer out of vad/mod.rs. Public gigastt_core::vad::* paths
are unchanged.
